Code modular and Tkinter out of class
